Create thumbnails module
